# lmdb: adds finalizer to unmanaged Txn objects, aborting them before gc